+++ CHANGELOG.md
@@ -4,6 +4,8 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Unreleased
 
+- Make `OrderingFilterBackend`/`SearchFilterBackend` show `action` with openapi warnings
+
 ## 0.1.0
 
 - Beta release